// angular/src/app/api/model/clienteFisicoResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EnderecoResponse } from './enderecoResponse';


/**
 * Dados completos de um cliente pessoa física
 */
export interface ClienteFisicoResponse { 
    /**
     * ID do cliente
     */
    readonly id?: number;
    /**
     * Tipo do cliente
     */
    readonly tipo?: ClienteFisicoResponse.TipoEnum;
    /**
     * CPF (formato: 000.000.000-00)
     */
    cpf?: string;
    /**
     * Nome completo
     */
    nome?: string;
    /**
     * RG (sem formatação)
     */
    rg?: string;
    /**
     * E-mail do cliente
     */
    email?: string;
    /**
     * Data de nascimento
     */
    dataNascimento?: string;
    /**
     * Status do cliente
     */
    estaAtivo?: boolean;
    /**
     * Lista de endereços do cliente
     */
    enderecos?: Array<EnderecoResponse>;
    /**
     * Data de criação
     */
    readonly createdAt?: string;
    /**
     * Data da última atualização
     */
    readonly updatedAt?: string;
}
export namespace ClienteFisicoResponse {
    export const TipoEnum = {
        Fisica: 'FISICA',
        Juridica: 'JURIDICA'
    } as const;
    export type TipoEnum = typeof TipoEnum[keyof typeof TipoEnum];
}